LayoutTests/http/tests/eventsource/eventsource-page-cache-connecting-expected.txt

Tests that a page with an EventSource that is connecting can enter the PageCache.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from cache
pagehide - entering cache
pageshow - from cache
PASS Page did enter and was restored from the page cache
PASS Received open event
PASS restoredFromPageCache is true
PASS Received message
PASS restoredFromPageCache is true
PASS receivedOpenEvent is true
PASS successfullyParsed is true

TEST COMPLETE
